Fixed profiler.base.PassThroughProfiler base class
