+++ docs/source/common/trainer.rst
@@ -1252,7 +1252,7 @@ See the :doc:`profiler documentation <../advanced/profiler>`. for more details.
 
 .. testcode::
 
-    from pytorch_lightning.profiler import SimpleProfiler, AdvancedProfiler
+    from pytorch_lightning.profilers import SimpleProfiler, AdvancedProfiler
 
     # default used by the Trainer
     trainer = Trainer(profiler=None)
