+++ pl_examples/basic_examples/profiler_example.py
@@ -31,7 +31,7 @@ import torchvision.transforms as T
 
 from pl_examples import _DATASETS_PATH, cli_lightning_logo
 from pytorch_lightning import LightningDataModule, LightningModule
-from pytorch_lightning.profiler.pytorch import PyTorchProfiler
+from pytorch_lightning.profilers.pytorch import PyTorchProfiler
 from pytorch_lightning.utilities.cli import LightningCLI
 
 DEFAULT_CMD_LINE = (
