+++ pytorch_lightning/profilers/__init__.py
@@ -1,0 +1,15 @@
+from pytorch_lightning.profilers.advanced import AdvancedProfiler
+from pytorch_lightning.profilers.base import AbstractProfiler, BaseProfiler, PassThroughProfiler
+from pytorch_lightning.profilers.pytorch import PyTorchProfiler
+from pytorch_lightning.profilers.simple import SimpleProfiler
+from pytorch_lightning.profilers.xla import XLAProfiler
+
+__all__ = [
+    "AbstractProfiler",
+    "BaseProfiler",
+    "AdvancedProfiler",
+    "PassThroughProfiler",
+    "PyTorchProfiler",
+    "SimpleProfiler",
+    "XLAProfiler",
+]
