+++ tests/profilers/test_xla_profiler.py
@@ -17,7 +17,7 @@ from multiprocessing import Event, Process
 import pytest
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.profiler import XLAProfiler
+from pytorch_lightning.profilers import XLAProfiler
 from pytorch_lightning.utilities import _TORCH_GREATER_EQUAL_1_8, _TPU_AVAILABLE
 from tests.helpers import BoringModel
 from tests.helpers.runif import RunIf
